dns: migrate initialization/cleanup to SocketDNS.c (Phase 2.6c)
Migrated initialization and cleanup functions from SocketDNS-internal.c
to SocketDNS.c as static functions. This completes Phase 2.6c of the DNS
unification effort.

Functions migrated:
- Initialization: allocate_dns_resolver(), initialize_dns_fields(),
  initialize_mutex(), initialize_synchronization(), initialize_pipe(),
  initialize_dns_components(), cleanup_on_init_failure()
- Cleanup: cleanup_pipe(), cleanup_mutex_cond(), drain_completion_pipe(),
  reset_dns_state(), destroy_dns_resources()
- Helper functions: create_completion_pipe(), set_pipe_nonblocking(),
  initialize_queue_condition(), initialize_result_condition()
- INIT_PTHREAD_PRIMITIVE macro

All functions maintain identical signatures and implementation.
Updated SocketDNS-private.h to remove extern declarations for migrated
functions.

Closes #1126
